Separate test file into multiple files
