+++ package.json
@@ -3,9 +3,14 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "jest": {
+    "testPathIgnorePatterns": [
+      "fixtures"
+    ]
+  },
   "scripts": {
-    "test": "jest --runInBand __tests__/*.test.js",
-    "watch": "jest --watch-all __tests__/*.test.js"
+    "test": "jest --runInBand",
+    "watch": "jest --watch-all"
   },
   "keywords": [],
   "author": "",
