+++ readme.md
@@ -38,6 +38,32 @@ npm run watch
 
 Follow the on-screen prompts to exit out of the constant runner.
 
+### Run test files individually
+
+There are a lot of tests that are contained in this project. You can run a single test file individually by putting the name of the file after `npm test`. You can even only put part of the file name.
+
+```
+npm test 01
+```
+
+Keep in mind that the testing framework, Jest, will attempt to match as many files as possible. So, for example, the following command will run multiple test files, because more than one test file includes the word `get`.
+
+```
+npm test get
+```
+
+### Run tests individually
+
+_After choosing a specific file to run,_ you can also specific which test you want to run, specifically. Add `.only` after either `test` or `describe` for the specific test you'd like to run.
+
+```js
+test.only("should return an array of everyone's name who is in the line, in order", () => {
+```
+
+This will either run the specific `test` or, in the case of adding `.only` to a `describe`, all of the tests for a specific function.
+
+> **NOTE:** Don't forget to remove this after you get the test to pass!
+
 ### Run file
 
 If you want to manually test out your file, you can do so by running the following command.
